Hopfield neural network
Hopfield neural network
 Updated 23 minutes ago
This code is for a typical hopfield neural network for python 3.7. It is possible to store a fix number of patterns and run the dynamics, but it also possible to put patterns along the way. I am currently trying to introduce the unlearning algorithm made by Plakhov, its humongous computation time(~O(N^3)) is worrying me, so any help is welcome. 
I think that the usual dynamics is sufficiently optimized, but I am not sure.